modif

+++ ft_printf/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By:  qcoudeyr <@student.42perpignan.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/27 10:50:34 by  qcoudeyr         #+#    #+#             */
-/*   Updated: 2023/04/11 13:34:15 by  qcoudeyr        ###   ########.fr       */
+/*   Updated: 2023/04/11 14:33:11 by  qcoudeyr        ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,7 +21,7 @@
 int		ft_printf(const char *str, ...);
 int		ft_putnbru(unsigned int n);
 void	*ft_memset(void *s, int c, size_t n);
-int		ft_putnbr_base(long int n, char *base);
+int		ft_putnbr_base(unsigned long long n, char *base);
 char	*ft_strrchr(const char *s, int c);
 char	*ft_strdup(const char *s);
 size_t	ft_strlen(const char *s);
